Update 'controls' components
- Clean up structure of components, rename variables
- Progress bar moved up betwenn scores and controls
- Some layout adjustments to fit smaller screens
