+++ src/components/game/controls/BoardSection.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { useDroppable } from '@dnd-kit/core'
 import { SortableContext, horizontalListSortingStrategy } from '@dnd-kit/sortable'
 
-import { type Dice } from './DnDDiceBoard'
+import { type Dice } from '../../../types'
 import DiceItem from './DiceItem'
 import SortableDiceItem from './SortableDiceItem'
 import styles from './DnDDiceBoard.module.sass'
