--- src/components/game/controls/DnDConstants.ts
@@ -1,6 +1,0 @@
-import { type DropAnimation } from '@dnd-kit/core'
-
-export const dropAnimation: DropAnimation = {
-  duration: 500,
-  easing: 'cubic-bezier(0.18, 0.67, 0.6, 1.22)'
-}
