+++ src/components/game/controls/DnDDiceBoard.module.sass
@@ -8,7 +8,7 @@
 .board
   display: flex
   border-radius: .25rem
-  min-width: 44px
+  min-width:20px
   background-color: $primary-light-color
   &:first-child
     background-color: $color-light-orange
@@ -16,13 +16,14 @@
     display: inherit
 
 .itemContainer
-  margin: 0 2px
+  margin: 0
 
 .item
   display: flex
-  min-width: 40px
+  min-width: 50px
   height: auto
   touch-action: none
+  margin: 0 .1rem
 
 .btnContainer
   flex-grow: 1
