+++ src/components/game/controls/DnDDiceBoard.tsx
@@ -14,132 +14,30 @@ import {
   DragOverlay
 } from '@dnd-kit/core'
 
-import { dropAnimation } from './DnDConstants'
-
 import { type RootState } from '../../../store'
 import { selectDice, deselectDice } from '../../../store/slices/shSlice'
+import type { Dice, Status, BoardSections } from '../../../types'
+import {
+  dropAnimation,
+  diceArray,
+  initializeBoard,
+  findBoardSectionContainer,
+  getDiceById
+} from './DnDHelpers'
 import BoardSection from './BoardSection'
 import DiceItem from './DiceItem'
 import MainButton from './MainButton'
 import styles from './DnDDiceBoard.module.sass'
 
-export type Status = 'roll' | 'sel'
-
-export interface Dice {
-  id: string
-  status: Status
-  value: number
-}
-
-export type BoardSections = Record<string, Dice[]>
-
-const diceArray: Dice[] = [
-  {
-    id: 'one',
-    status: 'roll',
-    value: 0
-  },
-  {
-    id: 'two',
-    status: 'roll',
-    value: 0
-  },
-  {
-    id: 'three',
-    status: 'roll',
-    value: 0
-  },
-  {
-    id: 'four',
-    status: 'roll',
-    value: 0
-  },
-  {
-    id: 'five',
-    status: 'roll',
-    value: 0
-  }
-]
-
-export const getDiceByStatus = (dice: Dice[], status: Status): Dice[] => {
-  return dice.filter((dice) => dice.status === status)
-}
-
-export const getDiceById = (dice: Dice[], id: string): Dice => {
-  return dice.find((dice) => dice.id === id) as Dice
-}
-
-export const BOARD_SECTIONS = {
-  sel: 'sel',
-  roll: 'roll'
-}
-
-export const initializeBoard = (dice: Dice[]): BoardSections => {
-  const boardSections: BoardSections = {}
-
-  Object.keys(BOARD_SECTIONS).forEach((boardSectionKey) => {
-    boardSections[boardSectionKey] = getDiceByStatus(
-      dice,
-      boardSectionKey as Status
-    )
-  })
-
-  return boardSections
-}
-
-export const findBoardSectionContainer = (
-  boardSections: BoardSections,
-  id: string
-): string => {
-  if (id in boardSections) {
-    return id
-  }
-
-  const container = Object.keys(boardSections).find((key) =>
-    boardSections[key].find((item) => item.id === id)
-  )
-  return container as string
-}
-
 const DnDDiceBoard: FC = () => {
+
   const { game } = useSelector((state: RootState) => state.sh)
   const dispatch = useDispatch()
+  const [diceState, setDiceState] = useState<Dice[]>(diceArray)
 
-  useEffect(() => {
-
-    const update = initialDice.filter(item => item.status === 'roll')
-    game.roll.forEach((value, index) => {
-      if (value > 0) {
-        update[index].value = value
-      }
-    })
-
-    const onHold = initialDice.filter(item => item.status === 'sel')
-    setCurrState([...onHold, ...update])
-  }, [game.roll])
-
-  useEffect(() => {
-
-    if (game.saved) {
-
-      diceArray.forEach((item, index) => {
-        diceArray[index].status = 'roll'
-        diceArray[index].value = 0
-      })
-      const init = initializeBoard([...diceArray])
-
-      setBoardSections(init)
-      setCurrState([...diceArray])
-    }
-
-  }, [game.saved])
-
-  const [initialDice, setCurrState] = useState<Dice[]>(diceArray)
-
-  const initialBoardSections = initializeBoard(initialDice)
+  const initialBoardSections = initializeBoard(diceState)
   const [boardSections, setBoardSections] =
     useState<BoardSections>(initialBoardSections)
-
   const [activeDiceId, setActiveDiceId] = useState<null | string>(null)
 
   const sensors = useSensors(
@@ -148,6 +46,9 @@ const DnDDiceBoard: FC = () => {
       coordinateGetter: sortableKeyboardCoordinates
     })
   )
+
+  // item to animate on interaction
+  const item = (activeDiceId != null) ? getDiceById(diceState, activeDiceId) : null
 
   const handleDragStart = ({ active }: DragStartEvent): void => {
     setActiveDiceId(active.id as string)
@@ -220,10 +121,10 @@ const DnDDiceBoard: FC = () => {
     }
 
     const activeIndex = boardSections[activeContainer].findIndex(
-      (task) => task.id === active.id
+      (dice) => dice.id === active.id
     )
     const overIndex = boardSections[overContainer].findIndex(
-      (task) => task.id === over?.id
+      (dice) => dice.id === over?.id
     )
 
     if (activeIndex !== overIndex) {
@@ -237,40 +138,64 @@ const DnDDiceBoard: FC = () => {
       }))
     }
 
-    const activeDice = getDiceById(initialDice, active.id.toString())
+    const activeDice = getDiceById(diceState, active.id.toString())
 
     if (activeDice.status !== activeContainer) {
       if (activeContainer === 'sel') { // !
         dispatch(selectDice(activeDice.value))
-        initialDice.forEach((item) => {
+        diceState.forEach((item) => {
           if (activeDice.id === item.id) {
             item.status = activeContainer as Status
           }
         })
-        setCurrState([...initialDice])
+        setDiceState([...diceState])
       } else {
-        initialDice.forEach((item) => {
+        diceState.forEach((item) => {
           if (activeDice.id === item.id) {
             item.status = activeContainer as Status
           }
         })
 
-        const selected = initialDice.filter(item => item.status === 'roll')
+        const selected = diceState.filter(item => item.status === 'roll')
 
         const data = {
           value: activeDice.value,
           order: selected.map(item => item.value)
         }
 
-        setCurrState([...initialDice])
-        dispatch(deselectDice(data)) // TODO: one thing!
+        setDiceState([...diceState])
+        dispatch(deselectDice(data))
       }
     }
 
     setActiveDiceId(null)
   }
 
-  const item = (activeDiceId != null) ? getDiceById(initialDice, activeDiceId) : null
+  useEffect(() => { // on roll
+    const update = diceState.filter(item => item.status === 'roll')
+    game.roll.forEach((value, index) => {
+      if (value > 0) {
+        update[index].value = value
+      }
+    })
+
+    const onHold = diceState.filter(item => item.status === 'sel')
+    setDiceState([...onHold, ...update])
+  }, [game.roll])
+
+  useEffect(() => { // on save
+    if (game.saved) {
+      diceArray.forEach((item, index) => {
+        diceArray[index].status = 'roll'
+        diceArray[index].value = 0
+      })
+
+      const init = initializeBoard([...diceArray])
+      setBoardSections(init)
+      setDiceState([...diceArray])
+    }
+
+  }, [game.saved])
 
   return (
     <div className={styles.container}>
@@ -290,6 +215,7 @@ const DnDDiceBoard: FC = () => {
             />
           </div>
         ))}
+        {/* draggable item animation */}
         <DragOverlay dropAnimation={dropAnimation}>
           {(item != null) ? <DiceItem dice={item} /> : null}
         </DragOverlay>
