+++ src/components/layout/ProgressBar.module.sass
@@ -2,6 +2,6 @@
 
 .progressBar
   background-color: $primary-color
-  margin-top: 1rem
+  margin: .5rem 0 1rem 0
   height: .1rem
   transition: width .5s ease-in
