+++ src/pages/Game.module.sass
@@ -6,13 +6,13 @@
   color: $primary-color
   h1
     text-align: center
-    margin: 1.5rem 0
+    margin: 1.25rem 0
     font-weight: 350
 
 .school
   display: grid
   grid-template-columns: repeat(6, 1fr)
-  gap: 1rem
+  gap: .75rem
   padding: 0 1rem
   svg
     color: $primary-color
@@ -27,7 +27,7 @@
 .schoolResult
   min-height: 2rem
   text-align: center
-  margin-bottom: 1rem
+  margin-bottom: .5rem
 
 .pre
   color: $color-orange
